+++ src/custom-snippets.ts
@@ -188,6 +188,7 @@ function interpolateCode(editor: vscode.TextEditor, code: string, context): stri
     .replace(/\$hover-line/g, context.hoverLine)
     .replace(/\$column/g, context.currentColumn)
     .replace(/\$hover-column/g, context.hoverColumn)
+    .replace(/\$file-text-str/g, JSON.stringify(context.currentFileText[1]))
     .replace(/\$file-text/g, context.currentFileText[1])
     .replace(/\$file/g, context.currentFilename?.replace(/\\/g, '\\\\') ?? '')
     .replace(/\$hover-file-text/g, context.hovercurrentFileText?.[1] ?? '')
@@ -196,20 +197,27 @@ function interpolateCode(editor: vscode.TextEditor, code: string, context): stri
     .replace(/\$editor-ns/g, context.editorNS)
     .replace(/\$repl/g, context.repl)
     .replace(/\$selection-closing-brackets/g, context.selectionWithBracketTrail?.[1])
+    .replace(/\$selection-str/g, JSON.stringify(context.selection))
     .replace(/\$selection/g, context.selection)
     .replace(/\$hover-text/g, context.hoverText);
   if (editor.document.languageId !== 'clojure') {
     return interpolateCode;
   } else {
     return interpolateCode
+      .replace(/\$current-form-str/g, JSON.stringify(context.currentForm[1]))
       .replace(/\$current-form/g, context.currentForm[1])
+      .replace(/\$current-pair-str/g, JSON.stringify(context.currentPair[1]))
       .replace(/\$current-pair/g, context.currentPair[1])
+      .replace(/\$enclosing-form-str/g, JSON.stringify(context.enclosingForm[1]))
       .replace(/\$enclosing-form/g, context.enclosingForm[1])
+      .replace(/\$top-level-form-str/g, JSON.stringify(context.topLevelForm[1]))
       .replace(/\$top-level-form/g, context.topLevelForm[1])
       .replace(/\$current-fn/g, context.currentFn[1])
       .replace(/\$top-level-fn/g, context.topLevelFn[1])
       .replace(/\$top-level-defined-symbol/g, context.topLevelDefinedForm?.[1] ?? '')
+      .replace(/\$head-str/g, JSON.stringify(context.head[1]))
       .replace(/\$head/g, context.head[1])
+      .replace(/\$tail-str/g, JSON.stringify(context.tail[1]))
       .replace(/\$tail/g, context.tail[1])
       .replace(/\$hover-current-form/g, context.hovercurrentForm?.[1] ?? '')
       .replace(/\$hover-current-pair/g, context.hovercurrentPair?.[1] ?? '')